requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
geopandas==1.1.3
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
pandas==3.0.1
pillow==12.1.1
Pygments==2.19.2
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.3.3
safetensors==0.7.0
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
timm==1.0.26
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.3
triangle==20250106
triton==3.3.1
typer==0.24.1
typing_extensions==4.15.0
wheel==0.46.3
